Ensure the windows changes work on linux.
