+++ component/prometheus/exporter/windows/config.go
@@ -4,74 +4,7 @@ import (
 	"strings"
 
 	windows_integration "github.com/grafana/agent/pkg/integrations/windows_exporter"
-	col "github.com/prometheus-community/windows_exporter/pkg/collector"
 )
-
-// DefaultArguments holds non-zero default options for Arguments when it is
-// unmarshaled from YAML.
-//
-// Some defaults are populated from init functions in the github.com/grafana/agent/pkg/integrations/windows_exporter package.
-
-var DefaultArguments = Arguments{
-	EnabledCollectors: strings.Split(windows_integration.DefaultConfig.EnabledCollectors, ","),
-	Dfsr: DfsrConfig{
-		SourcesEnabled: strings.Split(col.ConfigDefaults.Dfsr.DfsrEnabledCollectors, ","),
-	},
-	Exchange: ExchangeConfig{
-		EnabledList: strings.Split(col.ConfigDefaults.Exchange.CollectorsEnabled, ","),
-	},
-	IIS: IISConfig{
-		AppBlackList:  col.ConfigDefaults.Iis.AppExclude,
-		AppWhiteList:  col.ConfigDefaults.Iis.AppInclude,
-		SiteBlackList: col.ConfigDefaults.Iis.SiteExclude,
-		SiteWhiteList: col.ConfigDefaults.Iis.SiteInclude,
-		AppInclude:    col.ConfigDefaults.Iis.AppInclude,
-		AppExclude:    col.ConfigDefaults.Iis.AppExclude,
-		SiteInclude:   col.ConfigDefaults.Iis.SiteInclude,
-		SiteExclude:   col.ConfigDefaults.Iis.SiteExclude,
-	},
-	LogicalDisk: LogicalDiskConfig{
-		BlackList: col.ConfigDefaults.LogicalDisk.VolumeExclude,
-		WhiteList: col.ConfigDefaults.LogicalDisk.VolumeInclude,
-		Include:   col.ConfigDefaults.LogicalDisk.VolumeInclude,
-		Exclude:   col.ConfigDefaults.LogicalDisk.VolumeExclude,
-	},
-	MSMQ: MSMQConfig{
-		Where: col.ConfigDefaults.Msmq.QueryWhereClause,
-	},
-	MSSQL: MSSQLConfig{
-		EnabledClasses: strings.Split(col.ConfigDefaults.Mssql.EnabledCollectors, ","),
-	},
-	Network: NetworkConfig{
-		BlackList: col.ConfigDefaults.Net.NicExclude,
-		WhiteList: col.ConfigDefaults.Net.NicInclude,
-		Include:   col.ConfigDefaults.Net.NicInclude,
-		Exclude:   col.ConfigDefaults.Net.NicExclude,
-	},
-	Process: ProcessConfig{
-		BlackList: col.ConfigDefaults.Process.ProcessExclude,
-		WhiteList: col.ConfigDefaults.Process.ProcessInclude,
-		Include:   col.ConfigDefaults.Process.ProcessInclude,
-		Exclude:   col.ConfigDefaults.Process.ProcessExclude,
-	},
-	ScheduledTask: ScheduledTaskConfig{
-		Include: col.ConfigDefaults.ScheduledTask.TaskInclude,
-		Exclude: col.ConfigDefaults.ScheduledTask.TaskExclude,
-	},
-	Service: ServiceConfig{
-		UseApi: "false",
-		Where:  col.ConfigDefaults.Service.ServiceWhereClause,
-	},
-	SMTP: SMTPConfig{
-		BlackList: col.ConfigDefaults.Smtp.ServerExclude,
-		WhiteList: col.ConfigDefaults.Smtp.ServerInclude,
-		Include:   col.ConfigDefaults.Smtp.ServerInclude,
-		Exclude:   col.ConfigDefaults.Smtp.ServerExclude,
-	},
-	TextFile: TextFileConfig{
-		TextFileDirectory: col.ConfigDefaults.Textfile.TextFileDirectories,
-	},
-}
 
 // Arguments is used for controlling for this exporter.
 type Arguments struct {
@@ -91,11 +24,6 @@ type Arguments struct {
 	Service       ServiceConfig       `river:"service,block,optional"`
 	SMTP          SMTPConfig          `river:"smtp,block,optional"`
 	TextFile      TextFileConfig      `river:"text_file,block,optional"`
-}
-
-// SetToDefault implements river.Defaulter.
-func (a *Arguments) SetToDefault() {
-	*a = DefaultArguments
 }
 
 // Convert converts the component's Arguments to the integration's Config.
