+++ pkg/integrations/windows_exporter/config.go
@@ -5,70 +5,7 @@ import (
 	"github.com/grafana/agent/pkg/integrations"
 	integrations_v2 "github.com/grafana/agent/pkg/integrations/v2"
 	"github.com/grafana/agent/pkg/integrations/v2/metricsutils"
-	col "github.com/prometheus-community/windows_exporter/pkg/collector"
 )
-
-// DefaultConfig holds the default settings for the windows_exporter integration.
-var DefaultConfig = Config{
-	EnabledCollectors: "cpu,cs,logical_disk,net,os,service,system",
-	Dfsr: DfsrConfig{
-		SourcesEnabled: col.ConfigDefaults.Dfsr.DfsrEnabledCollectors,
-	},
-	Exchange: ExchangeConfig{
-		EnabledList: col.ConfigDefaults.Exchange.CollectorsEnabled,
-	},
-	IIS: IISConfig{
-		AppBlackList:  col.ConfigDefaults.Iis.AppExclude,
-		AppWhiteList:  col.ConfigDefaults.Iis.AppInclude,
-		SiteBlackList: col.ConfigDefaults.Iis.SiteExclude,
-		SiteWhiteList: col.ConfigDefaults.Iis.SiteInclude,
-		AppInclude:    col.ConfigDefaults.Iis.AppInclude,
-		AppExclude:    col.ConfigDefaults.Iis.AppExclude,
-		SiteInclude:   col.ConfigDefaults.Iis.SiteInclude,
-		SiteExclude:   col.ConfigDefaults.Iis.SiteExclude,
-	},
-	LogicalDisk: LogicalDiskConfig{
-		BlackList: col.ConfigDefaults.LogicalDisk.VolumeExclude,
-		WhiteList: col.ConfigDefaults.LogicalDisk.VolumeInclude,
-		Include:   col.ConfigDefaults.LogicalDisk.VolumeInclude,
-		Exclude:   col.ConfigDefaults.LogicalDisk.VolumeExclude,
-	},
-	MSMQ: MSMQConfig{
-		Where: col.ConfigDefaults.Msmq.QueryWhereClause,
-	},
-	MSSQL: MSSQLConfig{
-		EnabledClasses: col.ConfigDefaults.Mssql.EnabledCollectors,
-	},
-	Network: NetworkConfig{
-		BlackList: col.ConfigDefaults.Net.NicExclude,
-		WhiteList: col.ConfigDefaults.Net.NicInclude,
-		Include:   col.ConfigDefaults.Net.NicInclude,
-		Exclude:   col.ConfigDefaults.Net.NicExclude,
-	},
-	Process: ProcessConfig{
-		BlackList: col.ConfigDefaults.Process.ProcessExclude,
-		WhiteList: col.ConfigDefaults.Process.ProcessInclude,
-		Include:   col.ConfigDefaults.Process.ProcessInclude,
-		Exclude:   col.ConfigDefaults.Process.ProcessExclude,
-	},
-	ScheduledTask: ScheduledTaskConfig{
-		Include: col.ConfigDefaults.ScheduledTask.TaskInclude,
-		Exclude: col.ConfigDefaults.ScheduledTask.TaskExclude,
-	},
-	Service: ServiceConfig{
-		UseApi: "false",
-		Where:  col.ConfigDefaults.Service.ServiceWhereClause,
-	},
-	SMTP: SMTPConfig{
-		BlackList: col.ConfigDefaults.Smtp.ServerExclude,
-		WhiteList: col.ConfigDefaults.Smtp.ServerInclude,
-		Include:   col.ConfigDefaults.Smtp.ServerInclude,
-		Exclude:   col.ConfigDefaults.Smtp.ServerExclude,
-	},
-	TextFile: TextFileConfig{
-		TextFileDirectory: col.ConfigDefaults.Textfile.TextFileDirectories,
-	},
-}
 
 func init() {
 	integrations.RegisterIntegration(&Config{})
@@ -92,14 +29,6 @@ type Config struct {
 	MSMQ          MSMQConfig          `yaml:"msmq,omitempty"`
 	LogicalDisk   LogicalDiskConfig   `yaml:"logical_disk,omitempty"`
 	ScheduledTask ScheduledTaskConfig `yaml:"scheduled_task,omitempty"`
-}
-
-// UnmarshalYAML implements yaml.Unmarshaler for Config.
-func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	*c = DefaultConfig
-
-	type plain Config
-	return unmarshal((*plain)(c))
 }
 
 // Name returns the name used, "windows_explorer"
